% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello user fuction}
\usage{
hello(x)
}
\arguments{
\item{x}{the name of the user}
}
\value{
the output from \code{\link{print}}
}
\description{
Hello user fuction
}
\examples{
hello("Franklin")
\dontrun{
hello("john")
}
}
