% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printyFunk.R
\name{printyFunk}
\alias{printyFunk}
\title{PrintyFunk}
\usage{
printyFunk(x, y, z)
}
\arguments{
\item{z}{}
}
\value{
this returns \code{\link{print}}
}
\description{
PrintyFunk
}
\examples{
printyFunk(1, 2, 3)
\dontrun{
printyFunk(a, b, c)
}
}
