% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printyFunk.R
\name{printyFunk}
\alias{printyFunk}
\title{PrintyFunk}
\usage{
printyFunk(x, y)
}
\arguments{
\item{y}{}
}
\value{
returns a tibble
}
\description{
PrintyFunk
}
\examples{
printyFunk(x = rnorm(5), y = rnorm(5))
}
